chore: restructure graphs
