+++ apps/data/src/dags/assembly.py
@@ -1,0 +1,331 @@
+"""Hamilton graph for assembling the canonical Person record.
+
+The matching + position-assignment work is done upstream
+(`matching.py` picks the TIGER blockface; `geocode.py` assigns lat/lon
+for TIGER-matched and TIGER-miss voters). This module synthesizes
+those into the single canonical `persons_geocoded` record that
+downstream consumers query.
+
+Node dependency chain:
+
+    persons_best_match (matching.py)
+              │
+              │       refined_positions (geocode.py)     osm_only_matches (geocode.py)
+              │              │                                    │
+              └──────────────┴───────────┬────────────────────────┘
+                                         │
+                                canonical_addresses
+                                         │
+                                persons_geocoded ◄── persons_validated
+                                         │
+                                  geocoding_summary
+
+`canonical_addresses` picks the street name: OSM-derived if either OSM
+source matched, TIGER `full_name` otherwise (Option B — per-building
+identity wins for `building_id` stability).
+
+`persons_geocoded` is a pure assembly node: joins person fields,
+canonical address, coords, and (when available) TIGER match metadata.
+
+`geocoding_summary` reports match-rate diagnostics broken down by
+`position_source`.
+"""
+
+import duckdb
+from src.models import TableRef
+from src.tables import PERSON_CATALOG, ensure_org_schema, org_fqn
+
+
+def _current_version(conn: duckdb.DuckDBPyConnection) -> int:
+    return conn.sql(f"FROM {PERSON_CATALOG}.current_snapshot()").fetchone()[0]
+
+
+# ---------------------------------------------------------------------------
+# Node 1 – canonical address strings (street name resolved via OSM or TIGER)
+# ---------------------------------------------------------------------------
+
+
+def canonical_addresses(
+    persons_best_match: TableRef,
+    persons_decomposed: TableRef,
+    refined_positions: TableRef,
+    osm_only_matches: TableRef,
+    organization_slug: str,
+    conn: duckdb.DuckDBPyConnection,
+) -> TableRef:
+    """Per-person canonical (address_line_1, matched_tokens).
+
+    Street-name source priority — for each voter:
+      1. `refined_positions.osm_street` if non-null (the voter matched
+         an OSM building, regardless of whether they also matched
+         TIGER). Same OSM building → same canonical → same building_id.
+         This prevents the dupe pattern where TIGER-matched voters and
+         osm_only voters at the same physical building end up with
+         different canonical street names ("F D R Dr" vs "FDR Drive").
+      2. `osm_only_matches.osm_street` for TIGER-miss voters rescued
+         via direct OSM lookup.
+      3. `persons_best_match.full_name` (TIGER's authoritative form)
+         for tiger_only voters where no OSM building was found.
+
+    Half-coded addresses are preserved with the "1/2" in canonical
+    address_line_1.
+
+    matched_tokens is the tokenization of the canonical street via the
+    same scheme upstream tokenizers use.
+
+    Output schema: (external_id, address_line_1, matched_tokens).
+
+    Incremental: skips external_ids already present.
+    """
+    table_suffix = "canonical_addresses"
+    ensure_org_schema(conn, organization_slug)
+    fqn = org_fqn(organization_slug, table_suffix)
+    match_fqn = persons_best_match.fqn
+    decomposed_fqn = persons_decomposed.fqn
+    refined_fqn = refined_positions.fqn
+    osm_only_fqn = osm_only_matches.fqn
+
+    conn.execute(f"""
+        CREATE TABLE IF NOT EXISTS {fqn} (
+            external_id     VARCHAR,
+            address_line_1  VARCHAR,
+            matched_tokens  VARCHAR[]
+        )
+    """)
+
+    # For TIGER-matched voters, prefer OSM street when the OSM lookup
+    # hit (osm_street IS NOT NULL); fall back to TIGER full_name. For
+    # osm_only voters, always use OSM street.
+    conn.execute(f"""
+        INSERT INTO {fqn}
+        WITH src AS (
+            SELECT r.external_id,
+                   UPPER(COALESCE(r.osm_street, m.full_name)) AS street_canonical
+            FROM {refined_fqn} r
+            LEFT JOIN {match_fqn} m ON m.external_id = r.external_id
+            UNION ALL
+            SELECT external_id, UPPER(osm_street) AS street_canonical
+            FROM {osm_only_fqn}
+        )
+        SELECT
+            s.external_id,
+            COALESCE(d.house_num_prefix, '') || CAST(d.house_number AS VARCHAR)
+              || CASE WHEN d.half_code IS NOT NULL AND d.half_code != ''
+                      THEN ' ' || d.half_code
+                      ELSE '' END
+              || ' ' || s.street_canonical                    AS address_line_1,
+            list_distinct(list_filter(
+              list_concat(
+                regexp_split_to_array(lower(trim(s.street_canonical)), '[^a-z0-9]+'),
+                regexp_extract_all(lower(trim(s.street_canonical)), '[0-9]+')
+              ),
+              x -> length(x) > 0
+            ))                                                AS matched_tokens
+        FROM src s
+        INNER JOIN {decomposed_fqn} d ON d.external_id = s.external_id
+        WHERE s.external_id NOT IN (SELECT external_id FROM {fqn})
+    """)
+
+    version = _current_version(conn)
+    return TableRef(
+        catalog=PERSON_CATALOG,
+        schema=organization_slug,
+        table=table_suffix,
+        version=version,
+    )
+
+
+# ---------------------------------------------------------------------------
+# Node 2 – assemble the canonical geocoded persons table
+# ---------------------------------------------------------------------------
+
+
+def persons_geocoded(
+    persons_validated: TableRef,
+    persons_best_match: TableRef,
+    canonical_addresses: TableRef,
+    refined_positions: TableRef,
+    osm_only_matches: TableRef,
+    organization_slug: str,
+    conn: duckdb.DuckDBPyConnection,
+) -> TableRef:
+    """Canonical geocoded persons table — the single "person record" that
+    downstream consumers query against. Contains every voter that ended
+    up with coordinates, whether via TIGER blockface match or via
+    direct OSM lookup (`osm_only_matches`).
+
+    Pure assembly node: joins person fields from `persons_validated`,
+    canonical address from `canonical_addresses`, lat/lon from
+    `refined_positions` OR `osm_only_matches` (coalesced), TIGER
+    metadata from `persons_best_match` (LEFT JOIN — osm-only voters
+    legitimately have no row), and a `blockface_id` that's either the
+    TIGER blockface or the nearest-blockface snap from
+    `osm_only_matches`. `position_source` is `'osm_only'` for the
+    osm-only voters.
+
+    address_line_2 normalization (UPPER + TRIM) happens here because it
+    comes straight from `persons_validated` with no TIGER/OSM-derived
+    equivalent — SBOE has a small number of mixed-case "Num 1"-style
+    apartment-type rows. SBOE source data is otherwise already UPPER, so
+    line_1/city/state/etc. pass through.
+
+    Persons that didn't end up in either `refined_positions` or
+    `osm_only_matches` are excluded entirely — they live in
+    `persons_validated` only. Match-rate diagnostics in
+    `geocoding_summary` reconcile counts across both tables.
+
+    Address-derived stable keys (`building_id`, `door_id`) are computed
+    here once both `address_line_1` (canonical) and `address_line_2`
+    (validated) are in scope. See docs/product-model.md for the keying
+    convention. Single-family doors get a double-pipe in door_id (empty
+    middle segment) so building_id and door_id never collide.
+
+    Schema: drops and recreates on every run. Tolerates schema iteration
+    while the canonical-record shape is being settled.
+    """
+    table_suffix = "persons_geocoded"
+    ensure_org_schema(conn, organization_slug)
+    fqn = org_fqn(organization_slug, table_suffix)
+    persons_fqn = persons_validated.fqn
+    match_fqn = persons_best_match.fqn
+    canonical_fqn = canonical_addresses.fqn
+    coords_fqn = refined_positions.fqn
+    osm_only_fqn = osm_only_matches.fqn
+
+    conn.execute(f"DROP TABLE IF EXISTS {fqn}")
+    conn.execute(f"""
+        CREATE TABLE {fqn} AS
+        WITH normalized_persons AS (
+          SELECT
+              p.external_id,
+              p.external_id_type,
+              p.first_name,
+              p.last_name,
+              TRIM(UPPER(p.address_line_2)) AS address_line_2,
+              p.city,
+              p.state,
+              p.zip5,
+              p.zip4,
+              p.other_properties
+          FROM {persons_fqn} p
+        ),
+        -- Coordinates come from refined_positions for TIGER-matched
+        -- voters, osm_only_matches for the rule-3 rescues. Each voter
+        -- appears in exactly one (refined_positions excludes voters
+        -- without a TIGER blockface; osm_only_matches excludes those
+        -- with one).
+        positions AS (
+          SELECT external_id, latitude, longitude, position_source,
+                 NULL::VARCHAR AS osm_blockface_id
+          FROM {coords_fqn}
+          UNION ALL
+          SELECT external_id, latitude, longitude,
+                 'osm_only' AS position_source,
+                 blockface_id AS osm_blockface_id
+          FROM {osm_only_fqn}
+        )
+        SELECT
+            np.external_id,
+            np.external_id_type,
+            np.first_name,
+            np.last_name,
+            c.address_line_1,
+            np.address_line_2,
+            np.city,
+            np.state,
+            np.zip5,
+            np.zip4,
+            np.other_properties,
+            p.latitude,
+            p.longitude,
+            p.position_source,
+            COALESCE(m.blockface_id, p.osm_blockface_id)                    AS blockface_id,
+            m.person_house_number,
+            m.match_score,
+            (c.address_line_1 || '|' || np.zip5)                            AS building_id,
+            (c.address_line_1 || '|' || COALESCE(np.address_line_2, '') || '|' || np.zip5)
+                                                                            AS door_id
+        FROM normalized_persons np
+        INNER JOIN {canonical_fqn} c ON c.external_id = np.external_id
+        INNER JOIN positions p       ON p.external_id = np.external_id
+        LEFT  JOIN {match_fqn} m     ON m.external_id = np.external_id
+    """)
+
+    version = _current_version(conn)
+    return TableRef(
+        catalog=PERSON_CATALOG,
+        schema=organization_slug,
+        table=table_suffix,
+        version=version,
+    )
+
+
+# ---------------------------------------------------------------------------
+# Node 3 – summary diagnostics
+# ---------------------------------------------------------------------------
+
+
+def geocoding_summary(
+    persons_geocoded: TableRef,
+    persons_validated: TableRef,
+    organization_slug: str,
+    conn: duckdb.DuckDBPyConnection,
+) -> TableRef:
+    """Match-rate diagnostics: total comes from persons_validated (the
+    universal "all persons" table), matched comes from persons_geocoded
+    (only contains successfully-matched persons). Difference = unmatched.
+
+    Broken down by `position_source` so the TIGER pipeline vs OSM-only
+    rescue contributions are visible at a glance.
+
+    Always overwrites (non-incremental) since it is cheap and must reflect
+    the current state of both tables.
+    """
+    table_suffix = "geocoding_summary"
+    ensure_org_schema(conn, organization_slug)
+    fqn = org_fqn(organization_slug, table_suffix)
+    geocoded_fqn = persons_geocoded.fqn
+    persons_fqn = persons_validated.fqn
+
+    conn.execute(f"DROP TABLE IF EXISTS {fqn}")
+    conn.execute(f"""
+        CREATE TABLE {fqn} AS
+        WITH counts AS (
+            SELECT
+                (SELECT count(*) FROM {persons_fqn})                              AS total_persons,
+                (SELECT count(*) FROM {geocoded_fqn})                             AS matched,
+                (SELECT count(*) FROM {geocoded_fqn}
+                  WHERE position_source IN ('osm_matched','tiger_only','osm_complex'))
+                                                                                  AS matched_tiger,
+                (SELECT count(*) FROM {geocoded_fqn}
+                  WHERE position_source = 'osm_matched')                          AS matched_osm_road_projected,
+                (SELECT count(*) FROM {geocoded_fqn}
+                  WHERE position_source = 'osm_complex')                          AS matched_osm_complex,
+                (SELECT count(*) FROM {geocoded_fqn}
+                  WHERE position_source = 'tiger_only')                           AS matched_tiger_only,
+                (SELECT count(*) FROM {geocoded_fqn}
+                  WHERE position_source = 'osm_only')                             AS matched_osm_only
+        )
+        SELECT
+            total_persons,
+            matched,
+            total_persons - matched                                         AS unmatched,
+            round(100.0 * matched / NULLIF(total_persons, 0), 2)            AS match_pct,
+            matched_tiger,
+            matched_osm_road_projected,
+            matched_osm_complex,
+            matched_tiger_only,
+            matched_osm_only,
+            -- Backward-compat alias for older test/RPC consumers; counts every
+            -- voter that has any position (including osm_only).
+            matched                                                         AS blockface_matches
+        FROM counts
+    """)
+
+    version = _current_version(conn)
+    return TableRef(
+        catalog=PERSON_CATALOG,
+        schema=organization_slug,
+        table=table_suffix,
+        version=version,
+    )
